+++ test/blackbox-tests/test-cases/copy_files.t/run.t
@@ -10,3 +10,37 @@ Test that (copy_files ...) works
   Entering directory 'test2'
   # 1 "dummy.txt"
   hello
+
+Test (alias ...) and (mode ...) fields:
+
+  $ mkdir -p test3
+  $ cat >test3/dune-project <<EOF
+  > (lang dune 2.6)
+  > EOF
+  $ cat >test3/dune <<EOF
+  > (copy_files
+  >  (alias foo)
+  >  (mode promote-until-clean)
+  >  (files subdir/*.txt))
+  > EOF
+  $ mkdir -p test3/subdir
+  $ echo Foo >test3/subdir/foo.txt
+
+  $ dune build --root test3 @foo
+  Entering directory 'test3'
+  File "dune", line 2, characters 1-12:
+  2 |  (alias foo)
+       ^^^^^^^^^^^
+  Error: 'alias' is only available since version 2.7 of the dune language.
+  Please update your dune-project file to have (lang dune 2.7).
+  [1]
+
+  $ cat >test3/dune-project <<EOF
+  > (lang dune 2.7)
+  > EOF
+
+  $ dune build --root test3 @foo
+  Entering directory 'test3'
+
+  $ cat test3/foo.txt
+  Foo